requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
discord.py==2.2.2
exceptiongroup==1.1.1
fonttools==4.39.3
frozenlist==1.3.3
h11==0.14.0
idna==3.4
kiwisolver==1.4.4
lxml==4.9.2
matplotlib==3.6.3
multidict==6.0.4
mutagen==1.46.0
numpy==1.24.1
outcome==1.2.0
packaging==23.0
Pillow==9.5.0
pycparser==2.21
pycryptodomex==3.17
PyNaCl==1.5.0
pyOpenSSL==23.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.27.1
selenium==4.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
structlog==21.5.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.15
urllib3-secure-extra==0.1.0
websockets==11.0
wsproto==1.2.0
yarl==1.8.2
youtube-search==2.1.2
git+https://github.com/ytdl-org/youtube-dl.git@master#egg=youtube_dl
